+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,7 +1,11 @@
+## 0.2.35 (2024-6-6)
+
+- fix: load font family resources to user's system
+
 ## 0.2.34 (2024-6-6)
 
-- fix: ui optimizaion of setting container
-
+- feat: ui optimizaion of setting container
+  
 ## 0.2.33 (2024-6-6)
 
 - chore: update deps